05-project/midterm-poc/kube-owl/requirements.txt
fastapi
uvicorn[standard]
llama-index-core
llama-index-llms-ollama
llama-index-llms-gemini
llama-index-llms-anthropic
llama-index-embeddings-ollama
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
llama-index-readers-file
chromadb
python-multipart
python-dotenv
pydantic
pydantic-settings
paramiko
llama-index-llms-openai
